feat(EWT-1376): adds logic to create and wait for the SQL session
currently this async work is happening internal to the Connection constructor,
but going forward we might want to refactor it to some sort of static factory
that returns a Promise for a Connection instance so that async behavior
is more explicit.

brought in `fetch-retry` to manage the polling over http. this library
is probably overkill for this initial use case, but i think it will make
itself worthwhile when we later add in timeouts and retries in case of
unavailability errors

brought in `pino` for logging. it seems lightweight enough for the way
we're using it now as a debugging tool, but also something we can grow
into. i find its API a little awkward, so might also keep my eye out
for a replacement down the line.

i pinkie-promise the next PR will be unit tests