chore: add "type=module" to package.json
this requires all imports to have an explicit file extension, and
unintuitively that extension needs to be ".js" instead of ".ts" even
if the file is typescript, because the typescript transpiler will not
change it.

updating all local imports to use `@/` syntax was not a requirement,
just some cleanup because i was already touching every import statement.
and added a linter rule to keep things that way

had to switch to `tsx` for executing typescript files, `@swc-node` doesn't
support running ESM project code